refactor:put code from my_ciphering_cli to app
